autoHQ, minor fixes
- an automatic HQ/HD switch can now be toggled in /account.
- an overview menu was added to /account.
- an annotations/captions toggles were added to /account.
- fix video response thumbnails not loading
- fix standard mp4 being deleted when hd was done processing
- fix show more comments button not showing on viewmodel
- fix search options being over search suggestions.
(thanks, griffen!)
- html5 embeds no longer require an exp_hd parameter to show
hd button when video is cached

+++ back/yt2009utils.js
@@ -985,6 +985,15 @@ module.exports = {
             quality = "360p"
         }
 
+        if(yt2009exports.getStatus(`${id}-${quality}`)) {
+            yt2009exports.waitForStatusChange(`${id}-${quality}`, () => {
+                callback(`${id}-${quality}.mp4`)
+            })
+            return;
+        }
+
+        yt2009exports.updateFileDownload(`${id}-${quality}`, 1)
+
         fetch("https://www.youtube.com/youtubei/v1/player?key=AIzaSyAO_FJ2SlqU8Q4STEHLGCilw_Y9_11qcW8", {
             "headers": {
                 "accept": "*/*",
@@ -1017,6 +1026,7 @@ module.exports = {
             // parse formats
             if(!r.streamingData) {
                 callback(false)
+                yt2009exports.updateFileDownload(`${id}-${quality}`, 2)
                 return;
             }
             let qualities = {}
@@ -1072,6 +1082,7 @@ module.exports = {
             }
             if(!qualities[quality]) {
                 callback(false)
+                yt2009exports.updateFileDownload(`${id}-${quality}`, 2)
                 return;
             }
 
@@ -1109,11 +1120,14 @@ module.exports = {
                     setTimeout(() => {
                         // delete temp assets
                         try {
-                            fs.unlinkSync(`${__dirname}/${audioPath}`)
+                            if(!audioPath.includes(".mp4")) {
+                                fs.unlinkSync(`${__dirname}/${audioPath}`)
+                            }
                             fs.unlinkSync(`${__dirname}/${videoPath}`)
                         }
                         catch(error) {}
                     }, 10000)
+                    yt2009exports.updateFileDownload(`${id}-${quality}`, 2)
                 })
             }
         })})
